pyproject.toml
[project]
name = "browserist"
version = "1.7.16"
description = "Extension for the Selenium web driver that makes browser automation even easier"
authors = [
  { name = "Jakob Bagterp", email = "jakob_bagterp@hotmail.com" }
]
maintainers = [
  { name = "Jakob Bagterp", email = "jakob_bagterp@hotmail.com" }
]
license = { text = "MIT License" }
readme = "README.md"
keywords = ["python", "selenium", "browser", "automation"]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "lxml==6.0.2",
    "lxml-stubs==0.5.1",
    "pillow==12.1.1",
    "requests==2.32.5",
    "selenium==4.40.0",
    "types-requests==2.32.4.20260107",
]

[project.urls]
"Bug Tracker" = "https://github.com/jakob-bagterp/browserist/issues"
"Documentation" = "https://jakob-bagterp.github.io/browserist/"
"API Reference" = "https://jakob-bagterp.github.io/browserist/reference/"
"Source Code" = "https://github.com/jakob-bagterp/browserist"
"Release Notes" = "https://github.com/jakob-bagterp/browserist/releases"

[dependency-groups]
dev = [
    "coverage==7.13.2",
    "keyring==25.7.0",
    "mypy==1.19.1",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "pytest-cov==6.2.1",
    "pytest-xdist==3.8.0",
    "ruff==0.15.0",
    "tox==4.34.1",
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-git-revision-date-localized-plugin==1.5.1",
    "mkdocs-material==9.7.1",
    "mkdocs-material-extensions==1.3.1",
    "mkdocs-minify-plugin==0.8.0",
    "mkdocs-redirects==1.2.2",
    "mkdocstrings==1.0.2",
    "mkdocstrings-python==2.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/browserist"]

[tool.hatch.build.targets.sdist]
include = ["/src"]

[tool.hatch.build.targets.wheel.sources]
"src/browserist" = "browserist"

[tool.uv]
package = true

[tool.mypy]
files = ["src"]
disallow_any_unimported = true
no_implicit_reexport = true
show_error_codes = true
strict = true
warn_unreachable = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501", "COM812"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
